No longer use with-temp-buffer to pass match-data
The previous approach to passing match data to functions in
dap--launch-json-variables had a major issue: it changed the current buffer,
which meant that variables like "lineNumber" or "selectedText" could not work
properly. The new approach checks if there are capture groups, and passes the
variable to be expanded as a variable if there are. This retains the flexibility
of the old approach, in that functions that don't take any arguments can still
be used for variables without capture groups, while still allowing capture
groups to be used in dedicated functions. Thanks to this patch, all variables
now work correctly.